ai_mitra/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
anthropic==0.5.0
pydantic==1.10.8
python-dotenv==1.0.0
langdetect==1.0.9
httpx==0.24.1
